# openapi.yaml
openapi: 3.0.3
info:
  title: LeanAutoLinks REST API
  version: 1.0.0
  description: |
    The LeanAutoLinks REST API provides programmatic access to all plugin functionality:
    managing linking rules, monitoring the background processing queue, viewing applied
    links and statistics, managing exclusions, and monitoring performance and cache health.

    ## Authentication

    All endpoints require WordPress Application Passwords using HTTP Basic Auth.
    The authenticated user must have the `manage_options` capability.

    ```bash
    curl -u "admin:xxxx xxxx xxxx xxxx xxxx xxxx" \
      https://example.com/wp-json/leanautolinks/v1/rules
    ```

    Generate an Application Password at **Users > Profile > Application Passwords** in
    your WordPress admin dashboard.

    ## Pagination

    List endpoints follow WordPress REST API conventions. Use `page` and `per_page`
    query parameters to paginate results. Responses include the following headers:

    | Header | Description |
    |--------|-------------|
    | `X-WP-Total` | Total number of items matching the query |
    | `X-WP-TotalPages` | Total number of pages |

    ## Error Handling

    All error responses use a consistent JSON structure:

    ```json
    {
      "code": "invalid_parameter",
      "message": "The rule_type parameter must be one of: internal, affiliate, entity.",
      "data": { "status": 400, "param": "rule_type" }
    }
    ```

    ## Rate Limiting

...

    ## Versioning

...
  contact:
...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /rules:
    get:
      operationId: listRules
      summary: List all rules
      description: |
        Returns a paginated list of linking rules. You can filter by rule type
        and active status. Results are ordered by priority (ascending), then by
        creation date (descending).
      tags: [Rules]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: rule_type
          in: query
          description: Filter rules by type.
          required: false
          schema:
            $ref: '#/components/schemas/RuleType'
        - name: is_active
          in: query
          description: Filter by active status. Omit to return all rules regardless of status.
          required: false
          schema:
            type: integer
            enum: [0, 1]
        - name: search
          in: query
          description: Search rules by keyword (partial match).
          required: false
          schema:
            type: string
        - name: orderby
          in: query
          description: Field to order results by.
          required: false
          schema:
            type: string
            enum: [id, keyword, priority, created_at, updated_at]
            default: priority
        - name: order
          in: query
          description: Sort direction.
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: A paginated list of rules.
          headers:
            X-WP-Total:
              $ref: '#/components/headers/X-WP-Total'
            X-WP-TotalPages:
              $ref: '#/components/headers/X-WP-TotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
              example:
                - id: 1
                  rule_type: internal
                  keyword: inteligencia artificial
                  target_url: https://example.com/glosario/inteligencia-artificial/
                  entity_type: glossary
                  entity_id: 342
                  priority: 5
                  max_per_post: 1
                  case_sensitive: false
                  is_active: true
                  nofollow: false
                  sponsored: false
                  created_at: "2026-03-20T14:30:00"
                  updated_at: "2026-03-20T14:30:00"
                - id: 2
                  rule_type: affiliate
                  keyword: Notion
                  target_url: https://affiliate.example.com/notion?ref=lw
                  entity_type: null
                  entity_id: null
                  priority: 20
                  max_per_post: 1
                  case_sensitive: true
                  is_active: true
                  nofollow: true
                  sponsored: true
                  created_at: "2026-03-21T09:00:00"
                  updated_at: "2026-03-21T09:00:00"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createRule
      summary: Create a rule
      description: |
        Creates a new linking rule. After creation, the engine automatically
        enqueues affected posts (those containing the keyword) for background
        reprocessing.
      tags: [Rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
            examples:
              glossary_rule:
                summary: Glossary term rule
                value:
                  rule_type: internal
                  keyword: inteligencia artificial
                  target_url: https://example.com/glosario/inteligencia-artificial/
                  entity_type: glossary
                  entity_id: 342
                  priority: 5
                  max_per_post: 1
                  case_sensitive: false
                  nofollow: false
                  sponsored: false
              affiliate_rule:
                summary: Affiliate product rule
                value:
                  rule_type: affiliate
                  keyword: Notion
                  target_url: https://affiliate.example.com/notion?ref=lw
                  priority: 20
                  max_per_post: 1
                  case_sensitive: true
                  nofollow: true
                  sponsored: true
              entity_rule:
                summary: Entity/actor rule
                value:
                  rule_type: entity
                  keyword: Y Combinator
                  target_url: https://example.com/actores/y-combinator/
                  entity_type: actor
                  entity_id: 87
                  priority: 10
                  max_per_post: 1
      responses:
        '201':
          description: Rule created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /rules/import:
    post:
      operationId: importRules
      summary: Bulk import rules
      description: |
        Import multiple rules at once from a JSON array or CSV file. Each rule
        in the payload is validated individually. Successfully imported rules
        trigger reprocessing of affected posts. The response includes counts
        of created, skipped (duplicate keyword), and failed entries.
      tags: [Rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rules]
              properties:
                rules:
                  type: array
                  description: Array of rule objects to import.
                  minItems: 1
                  maxItems: 5000
                  items:
                    $ref: '#/components/schemas/RuleCreate'
                skip_duplicates:
                  type: boolean
                  description: |
                    When true, rules with a keyword that already exists are
                    silently skipped. When false, duplicate keywords cause
                    a validation error for that entry.
                  default: true
            example:
              skip_duplicates: true
              rules:
                - rule_type: internal
                  keyword: machine learning
                  target_url: https://example.com/glosario/machine-learning/
                  priority: 5
                - rule_type: entity
                  keyword: Sequoia Capital
                  target_url: https://example.com/actores/sequoia-capital/
                  entity_type: actor
                  entity_id: 112
                  priority: 10
          text/csv:
            schema:
              type: string
              description: |
                CSV file with headers matching rule fields. Required columns:
                `rule_type`, `keyword`, `target_url`. Optional columns:
                `entity_type`, `entity_id`, `priority`, `max_per_post`,
                `case_sensitive`, `nofollow`, `sponsored`.
              example: |
                rule_type,keyword,target_url,priority,max_per_post,nofollow,sponsored
                internal,machine learning,https://example.com/glosario/machine-learning/,5,1,0,0
                affiliate,Notion,https://affiliate.example.com/notion?ref=lw,20,1,1,1
      responses:
        '200':
          description: Import completed. Check the counts for details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                    description: Number of rules successfully created.
                    example: 48
                  skipped:
                    type: integer
                    description: Number of rules skipped due to duplicate keywords.
                    example: 2
                  failed:
                    type: integer
                    description: Number of rules that failed validation.
                    example: 0
                  errors:
                    type: array
                    description: Validation error details for failed entries.
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                          description: Zero-based index of the failed entry in the input array.
                        keyword:
                          type: string
                          description: The keyword of the failed entry (if parseable).
                        message:
                          type: string
                          description: Human-readable error message.
                    example: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---------------------------------------------------------------------------
  # Queue
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        WordPress Application Passwords. The username is your WordPress
        username and the password is an Application Password generated at
        **Users > Profile > Application Passwords**.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Current page of the collection.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Maximum number of items to return per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  headers:
    X-WP-Total:
      description: Total number of items in the collection.
      schema:
        type: integer
    X-WP-TotalPages:
      description: Total number of pages in the collection.
      schema:
        type: integer

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    RuleType:
      type: string
      enum: [internal, affiliate, entity]
      description: |
        The type of linking rule:
        - `internal`: Glossary terms and internal content links (TTL: 24h).
        - `affiliate`: Affiliate product/service links with `rel="sponsored nofollow"` (TTL: 12h).
        - `entity`: Entity/actor links for companies, VCs, and people (TTL: 24h).
...
    Rule:
      type: object
      description: A linking rule that maps a keyword to a target URL.
      required:
        - id
        - rule_type
        - keyword
        - target_url
        - priority
        - max_per_post
        - case_sensitive
        - is_active
        - nofollow
        - sponsored
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique rule identifier.
          readOnly: true
          example: 1
        rule_type:
          $ref: '#/components/schemas/RuleType'
        keyword:
          type: string
          maxLength: 255
          description: |
            The keyword or phrase to match in post content. Matched using
            word-boundary-aware regex (see matching algorithm documentation).
          example: inteligencia artificial
        target_url:
          type: string
          format: uri
          description: The URL that the matched keyword links to.
          example: https://example.com/glosario/inteligencia-artificial/
        entity_type:
          type: string
          maxLength: 100
          nullable: true
          description: |
            The WordPress custom post type associated with this rule (e.g.,
            `glossary`, `actor`). Null for affiliate rules or rules not tied
            to a CPT.
          example: glossary
        entity_id:
          type: integer
          format: int64
          nullable: true
          description: |
            The WordPress post ID of the associated entity. Null for rules
            not tied to a specific post.
          example: 342
        priority:
          type: integer
          minimum: 1
          maximum: 127
          description: |
            Processing priority. Lower values are processed first. When
            multiple rules match overlapping text in the same post, the
            higher-priority (lower value) rule wins.
          default: 10
          example: 5
        max_per_post:
          type: integer
          minimum: 1
          maximum: 127
          description: |
            Maximum number of times this rule can be applied within a single
            post. Prevents over-linking for high-frequency keywords.
          default: 1
          example: 1
        case_sensitive:
          type: boolean
          description: |
            Whether keyword matching is case-sensitive. Default is false
            (case-insensitive), which is recommended for most use cases.
          default: false
          example: false
        is_active:
          type: boolean
          description: Whether this rule is currently active. Inactive rules are not evaluated by the engine.
          default: true
          example: true
        nofollow:
          type: boolean
          description: |
            Whether to add `rel="nofollow"` to the generated link. Automatically
            set to true for affiliate rules.
          default: false
          example: false
        sponsored:
          type: boolean
          description: |
            Whether to add `rel="sponsored"` to the generated link. Automatically
            set to true for affiliate rules. Affiliate links always render as
            `rel="sponsored nofollow"`.
          default: false
          example: false
        created_at:
          type: string
          format: date-time
          description: When the rule was created.
          readOnly: true
          example: "2026-03-20T14:30:00"
        updated_at:
          type: string
          format: date-time
          description: When the rule was last updated.
          readOnly: true
          example: "2026-03-20T14:30:00"
    RuleCreate:
      type: object
      description: Request body for creating a new rule.
      required:
        - rule_type
        - keyword
        - target_url
      properties:
        rule_type:
          $ref: '#/components/schemas/RuleType'
        keyword:
          type: string
          maxLength: 255
          minLength: 1
          description: The keyword or phrase to match in post content.
        target_url:
          type: string
          format: uri
          description: The URL that the matched keyword links to.
        entity_type:
          type: string
          maxLength: 100
          nullable: true
          description: The associated custom post type.
        entity_id:
          type: integer
          format: int64
          nullable: true
          description: The associated entity post ID.
        priority:
          type: integer
          minimum: 1
          maximum: 127
          default: 10
        max_per_post:
          type: integer
          minimum: 1
          maximum: 127
          default: 1
        case_sensitive:
          type: boolean
          default: false
        nofollow:
          type: boolean
          default: false
        sponsored:
          type: boolean
          default: false
...
    Error:
      type: object
      description: |
        Standard error response following WordPress REST API conventions.
      required:
        - code
        - message
        - data
      properties:
        code:
          type: string
          description: |
            Machine-readable error code. Common codes:
            - `rest_forbidden`: Missing required capability.
            - `rest_no_route`: Endpoint not found.
            - `invalid_parameter`: A request parameter failed validation.
            - `missing_parameter`: A required parameter is missing.
            - `not_found`: The requested resource does not exist.
            - `duplicate_keyword`: A rule with this keyword already exists.
            - `internal_error`: An unexpected server error occurred.
          example: invalid_parameter
        message:
          type: string
          description: Human-readable error description.
          example: "The rule_type parameter must be one of: internal, affiliate, entity."
        data:
          type: object
          description: Additional error context.
          properties:
            status:
              type: integer
              description: The HTTP status code.
              example: 400
            param:
              type: string
              description: The parameter that caused the error (if applicable).
              example: rule_type
            details:
              type: object
              description: Additional error details (structure varies by error type).

    # --- Pagination ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: |
        The request is malformed or contains invalid parameters. Check the
        `code` and `data.param` fields to identify the issue.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: invalid_parameter
            message: "The rule_type parameter must be one of: internal, affiliate, entity."
            data:
              status: 400
              param: rule_type
    Unauthorized:
      description: |
        Authentication is missing or invalid. Ensure you are sending a valid
        WordPress Application Password via HTTP Basic Auth.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rest_not_logged_in
            message: "You are not currently logged in."
            data:
              status: 401
    Forbidden:
      description: |
        The authenticated user does not have the `manage_options` capability
        required to access this endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rest_forbidden
            message: "Sorry, you are not allowed to manage LeanAutoLinks settings."
            data:
              status: 403
...
    UnprocessableEntity:
      description: |
        The request body is syntactically valid but contains semantic errors.
        For example, a required field is missing or a value violates business rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: missing_parameter
            message: "The keyword field is required."
            data:
              status: 422
              param: keyword
    InternalError:
      description: An unexpected server error occurred. Check the WordPress error log for details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: "An unexpected error occurred. Please try again later."
            data:
              status: 500